+++ packages/components/organisms/f-content-cards/test-utils/component-objects/f-content-cards-home-promotion.selectors.js
@@ -1,11 +1,24 @@
-exports.PROMOTION_CARD_1_COMPONENT = '[data-test-id="home-promotion-1"]';
-exports.INNER_CONTAINER = '[data-test-id="home-promotion-1--container"]';
-exports.HEADER = '[data-test-id="home-promotion-1--header"]';
-exports.IMAGE = '[data-test-id="home-promotion-1--image"]';
-exports.SUBTITLE = '[data-test-id="home-promotion-1--subtitle"]';
+const PROMOTION_CARD_1_COMPONENT = '[data-test-id="home-promotion-1"]';
+const INNER_CONTAINER = '[data-test-id="home-promotion-1--container"]';
+const HEADER = '[data-test-id="home-promotion-1--header"]';
+const IMAGE = '[data-test-id="home-promotion-1--image"]';
+const SUBTITLE = '[data-test-id="home-promotion-1--subtitle"]';
 
-exports.PROMOTION_CARD_2_COMPONENT = '[data-test-id="home-promotion-2"]';
-exports.CTA = '[data-test-id="home-promotion-2--cta"]';
-exports.TEXT = '[data-test-id="home-promotion-2--text"]';
-exports.TITLE = '[data-test-id="home-promotion-2--title"]';
-exports.BACKGROUND_IMAGE = '[data-test-id="home-promotion-2--backgroundImage"]';
+const PROMOTION_CARD_2_COMPONENT = '[data-test-id="home-promotion-2"]';
+const CTA = '[data-test-id="home-promotion-2--cta"]';
+const TEXT = '[data-test-id="home-promotion-2--text"]';
+const TITLE = '[data-test-id="home-promotion-2--title"]';
+const BACKGROUND_IMAGE = '[data-test-id="home-promotion-2--backgroundImage"]';
+
+export {
+    PROMOTION_CARD_1_COMPONENT,
+    INNER_CONTAINER,
+    HEADER,
+    IMAGE,
+    SUBTITLE,
+    PROMOTION_CARD_2_COMPONENT,
+    CTA,
+    TEXT,
+    TITLE,
+    BACKGROUND_IMAGE
+};