+++ packages/components/organisms/f-content-cards/test/component/f-content-cards-home-promotion-card2.component.spec.js
@@ -1,28 +1,25 @@
-const HomePromotionCard2 = require('../../test-utils/component-objects/f-content-cards-home-promotion-card2.component');
-
-let card;
+import HomePromotionCard from '../../test-utils/component-objects/f-content-cards-home-promotion-card2.component';
 
 describe('home promotion card 2 component tests', () => {
     beforeEach(async () => {
-        card = new HomePromotionCard2();
-        await card.load();
+        await HomePromotionCard.load();
     });
 
     it('should display the card', async () => {
         // Assert
-        await expect(await card.isComponentDisplayed()).toBe(true);
+        await expect(await HomePromotionCard.isComponentDisplayed()).toBe(true);
     });
 
     it('should display the elements of the card', async () => {
         // Assert
-        await expect(await card.isCallToActionDisplayed()).toBe(true);
-        await expect(await card.isTextDisplayed()).toBe(true);
-        await expect(await card.isTitleDisplayed()).toBe(true);
-        await expect(await card.isBackgroundImageDisplayed()).toBe(true);
+        await expect(await HomePromotionCard.isCallToActionDisplayed()).toBe(true);
+        await expect(await HomePromotionCard.isTextDisplayed()).toBe(true);
+        await expect(await HomePromotionCard.isTitleDisplayed()).toBe(true);
+        await expect(await HomePromotionCard.isBackgroundImageDisplayed()).toBe(true);
     });
 
     it('should check that the card is clickable', async () => {
         // Assert
-        await expect(await card.isComponentClickable()).toBe(true);
+        await expect(await HomePromotionCard.isComponentClickable()).toBe(true);
     });
 });