+++ packages/components/organisms/f-footer/test/accessibility/axe-accessibility.spec.js
@@ -1,32 +1,43 @@
-import forEach from 'mocha-each';
+import Footer from '../../test-utils/component-objects/f-footer.component';
 
-const { getAxeResults } = require('../../../../../../test/utils/axe-helper');
-const Footer = require('../../test-utils/component-objects/f-footer.component');
-
-let footer;
-
+let locales;
 describe('Accessibility tests', () => {
-    beforeEach(() => {
-        footer = new Footer();
+    locales = [
+        'en-GB',
+        'en-AU',
+        'en-IE',
+        'en-NZ',
+        'es-ES',
+        'it-IT'
+    ];
+
+    locales.forEach(locale => {
+        it(`a11y - should test f-footer component WCAG compliance for country code "${locale}" with default options selected`, async () => {
+            // Act
+            await Footer.load({ locale });
+            const axeResults = await Footer.getAxeResults('f-footer');
+
+            // Assert
+            expect(axeResults.violations.length).toBe(0);
+        });
     });
 
-    forEach(['en-GB', 'en-AU', 'en-IE', 'en-NZ', 'es-ES', 'it-IT'])
-        .it('a11y - should test f-footer component WCAG compliance for country code "%s" with default options selected', tenant => {
+
+    locales = [
+        'en-GB',
+        'en-AU',
+        'en-IE',
+        'en-NZ'
+    ];
+
+    locales.forEach(locale => {
+        it(`a11y - should test f-footer component WCAG compliance for country code "${locale}" with extra options selected`, async () => {
             // Act
-            footer.load({ locale: tenant });
-            const axeResults = getAxeResults('f-footer');
+            await Footer.load({ locale, showCountrySelector: true });
+            const axeResults = await Footer.getAxeResults('f-footer');
 
             // Assert
             expect(axeResults.violations.length).toBe(0);
         });
-
-    forEach(['en-GB', 'en-AU', 'en-IE', 'en-NZ'])
-        .it('a11y - should test f-footer component WCAG compliance for country code "%s" with extra options selected', tenant => {
-            // Act
-            footer.load({ locale: tenant, showCountrySelector: true });
-            const axeResults = getAxeResults('f-footer');
-
-            // Assert
-            expect(axeResults.violations.length).toBe(0);
-        });
+    });
 });