+++ packages/components/organisms/f-header/test/visual/f-header.visual.desktop.spec.js
@@ -1,82 +1,90 @@
-import forEach from 'mocha-each';
+import Header from '../../test-utils/component-objects/f-header.component';
 
-const Header = require('../../test-utils/component-objects/f-header.component');
-
-let header;
+let tests;
 
 describe('Shared - f-header component tests', () => {
-    beforeEach(() => {
-        // Arrange
-        header = new Header();
+    tests = [
+        { tenant: 'en-GB', isLoggedIn: true },
+        { tenant: 'en-GB', isLoggedIn: false },
+        { tenant: 'en-AU', isLoggedIn: true },
+        { tenant: 'en-AU', isLoggedIn: false },
+        { tenant: 'en-NZ', isLoggedIn: true },
+        { tenant: 'en-NZ', isLoggedIn: false },
+        { tenant: 'en-IE', isLoggedIn: true },
+        { tenant: 'en-IE', isLoggedIn: false },
+        { tenant: 'it-IT', isLoggedIn: true },
+        { tenant: 'it-IT', isLoggedIn: false },
+        { tenant: 'es-ES', isLoggedIn: true },
+        { tenant: 'es-ES', isLoggedIn: false }
+    ];
+
+    tests.forEach(({ tenant, isLoggedIn }) => {
+        it('should display component', async () => {
+            // Arrange
+            const querystring = {
+                locale: tenant,
+                showOffersLink: true,
+                showDeliveryEnquiry: true
+            };
+
+            if (!isLoggedIn) {
+                querystring.userInfoProp = '!undefined';
+            }
+
+            // Act
+            await Header.load(querystring);
+
+            // Assert
+            await browser.percyScreenshot(`f-header - Base state - isLoggedIn: ${isLoggedIn} - ${tenant}`, 'desktop');
+        });
     });
 
-    forEach([
-        ['en-GB', true],
-        ['en-GB', false],
-        ['en-AU', true],
-        ['en-AU', false],
-        ['en-NZ', true],
-        ['en-NZ', false],
-        ['en-IE', true],
-        ['en-IE', false],
-        ['it-IT', true],
-        ['it-IT', false],
-        ['es-ES', true],
-        ['es-ES', false]
-    ]).it('should display component', async (tenant, isLoggedIn) => {
-        // Arrange
-        const querystring = {
-            locale: tenant,
-            showOffersLink: true,
-            showDeliveryEnquiry: true
-        };
+    tests = [
+        { tenant: 'en-GB', theme: 'white' },
+        { tenant: 'en-GB', theme: 'highlight' },
+        { tenant: 'en-GB', theme: 'transparent' }
+    ];
 
-        if (!isLoggedIn) {
-            querystring.userInfoProp = '!undefined';
-        }
+    tests.forEach(({ tenant, theme }) => {
+        it(`should display the "${tenant} - ${theme}" header theme`, async () => {
+            // Act
+            await Header.load({ locale: tenant, headerBackgroundTheme: `${theme}` });
 
-        // Act
-        await header.load(querystring);
-
-        // Assert
-        await browser.percyScreenshot(`f-header - Base state - isLoggedIn: ${isLoggedIn} - ${tenant}`, 'desktop');
+            // Assert
+            await browser.percyScreenshot(`f-header - Theme colours - ${theme}`, 'desktop');
+        });
     });
 
-    forEach([
-        ['en-GB', 'white'],
-        ['en-GB', 'highlight'],
-        ['en-GB', 'transparent']
-    ]).it('should display the "%s - %s" header theme', async (tenant, theme) => {
-        // Act
-        await header.load({ locale: tenant, headerBackgroundTheme: `${theme}` });
+    tests = [
+        { tenant: 'en-GB', control: 'showLoginInfo' },
+        { tenant: 'en-GB', control: 'showHelpLink' },
+        { tenant: 'en-GB', control: 'showCountrySelector' }
+    ];
 
-        // Assert
-        await browser.percyScreenshot(`f-header - Theme colours - ${theme}`, 'desktop');
+    tests.forEach(({ tenant, control }) => {
+        it(`should not display "${control}" `, async () => {
+            // Act
+            await Header.load({ locale: tenant, [control]: false });
+
+            // Assert
+            await browser.percyScreenshot(`f-header - ${control} - False`, 'desktop');
+        });
     });
 
-    forEach([
-        ['en-GB', 'showLoginInfo'],
-        ['en-GB', 'showHelpLink'],
-        ['en-GB', 'showCountrySelector']
-    ]).it('should not display "%s" ', async (tenant, control) => {
-        // Act
-        await header.load({ locale: tenant, [control]: false });
+    tests = [
+        { tenant: 'en-GB', link: 'userAccount' },
+        { tenant: 'en-GB', link: 'countrySelector' }
+    ];
 
-        // Assert
-        await browser.percyScreenshot(`f-header - ${control} - False`, 'desktop');
-    });
+    tests.forEach(({ tenant, link }) => {
+        it(`should display the "${tenant} ${link}" dropdown lists on hover`, async () => {
+            // Act
+            await Header.load({ locale: tenant });
+            await Header.moveToNavigationLink(link);
 
-    forEach([
-        ['en-GB', 'userAccount'],
-        ['en-GB', 'countrySelector']
-    ])
-    .it('should display the "%s %s" dropdown lists on hover', async (tenant, link) => {
-        // Act
-        await header.load({ locale: tenant });
-        await header.moveToNavigationLink(link);
-
-        // Assert
-        await browser.percyScreenshot('f-header - %s dropdown', 'desktop');
+            // Assert
+            await browser.percyScreenshot(`f-header - ${link} dropdown`, 'desktop');
+        });
     });
 
     // Not currently possible to set complex values (i.e., arrays) for controls via query strings.
@@ -84,13 +92,13 @@ describe('Shared - f-header component tests', () => {
     // https://github.com/storybookjs/storybook/issues/14420
     it.skip('should display any custom links', async () => {
         // Act
-        await header.load({
+        await Header.load({
             locale: 'en-GB',
             showOffersLink: false,
             showDeliveryEnquiry: false
             // Set custom links here
         });
-        await header.openMobileNavigationBar();
+        await Header.openMobileNavigationBar();
 
         // Assert
         await browser.percyScreenshot('f-header - with custom nav links', 'desktop');
@@ -101,7 +109,7 @@ describe('Shared - f-header component tests', () => {
     // https://github.com/storybookjs/storybook/issues/14420
     it.skip('should be able to show only custom links', async () => {
         // Act
-        await header.load({
+        await Header.load({
             locale: 'en-GB',
             showHelpLink: false,
             showLoginInfo: false,
@@ -110,7 +118,7 @@ describe('Shared - f-header component tests', () => {
             showDeliveryEnquiry: false
             // Set custom links here
         });
-        await header.openMobileNavigationBar();
+        await Header.openMobileNavigationBar();
 
         // Assert
         await browser.percyScreenshot('f-header - custom nav links only', 'desktop');
@@ -118,8 +126,8 @@ describe('Shared - f-header component tests', () => {
 
     it('should display link in hover state', async () => {
         // Act
-        await header.load();
-        await header.hoverOverLink('help');
+        await Header.load();
+        await Header.hoverOverLink('help');
 
         // Assert
         await browser.percyScreenshot('f-header - hover state', 'desktop');
@@ -127,8 +135,8 @@ describe('Shared - f-header component tests', () => {
 
     it('should display link in focus state', async () => {
         // Act
-        await header.load();
-        await header.pressDownTabKey(3);
+        await Header.load();
+        await Header.pressDownTabKey(3);
 
         // Assert
         await browser.percyScreenshot('f-header - focus state', 'desktop');